ECS Part 9 | ModelRenderer (#2682)
Ninth step of splittig up the 3000 lines big PR:
https://github.com/PixelGuys/Cubyz/pull/2652
(requires Part 8)

**QUANTUM DON'T BE SCARED OF THE BIG SCARY LOC NUMBER!**
**THIS PR FOR THE MOST PART ONLY MOVES CODE!**

``entity_manager.zig`` => ``modelRenderer.zig``

---------

+++ src/gui/windows/players.zig
@@ -21,7 +21,7 @@ pub var window = GuiWindow{
 
 const padding: f32 = 8;
 var userList: []*main.server.User = &.{};
-var entityCount: u32 = 0;
+var entityCount: usize = 0;
 
 fn kickbyConnection(conn: *main.network.Connection) void {
 	conn.disconnect();
